aipc_examples/document_classification_aipc/bert/implementation/src/requirements.txt
transformers==4.26.1
scikit-learn==1.2.2
scikit-multilearn==0.2.0
numpy==1.23.4
lsg-converter==0.0.5
sentence-transformers==2.2.2
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
compress_fasttext==0.1.3
scipy==1.10.0
nltk==3.8.1
gensim==4.3.0
ufal.udpipe==1.3.0.1
pyyaml==6.0
stop-words==2018.7.23
spacy==3.5.1
PageRange==0.4
jsonargparse[signatures]  # CLI
torch
torchvision==0.17.0
